refactor: extract shared TestFixtureBase for spawnMob/loginPlayer (#359)
Extract a shared TestFixtureBase interface that CombatTestFixture,
AbilityTestFixture, and StatusEffectTestFixture all implement.
This eliminates ~130 lines of duplicated spawnMob() and loginPlayer()
methods across the three fixture classes. The inline spawnMob extension
in StatusEffectSystemTest is also removed in favor of the inherited
method.
